+++ src/lib.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::must_use_candidate)]
+
 mod cell_widths;
 mod emoji_codes;
 mod emoji_replace;
@@ -6,4 +8,5 @@ mod filesize;
 mod ratio;
 mod region;
 
+pub mod cells;
 pub mod color_triplet;