+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+
 - Port `_emoji_codes.py` (`emoji_codes.rs`) in [#64](https://github.com/wizard-28/wealthy/pull/64) from [@wizard-28](https://github.com/wizard-28)
 - Port `_emoji_replace.py` (`emoji_replace.rs`) in [#64](https://github.com/wizard-28/wealthy/pull/64) from [@wizard-28](https://github.com/wizard-28)
 - Port `_ratio.py` (`ratio.rs`) in [#67](https://github.com/wizard-28/wealthy/pull/67) from [@wizard-28](https://github.com/wizard-28)
@@ -13,3 +16,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Port `color_triplet.py` (`color_triplet.rs`) in [#78](https://github.com/wizard-28/wealthy/pull/78) from [@wizard-28](https://github.com/wizard-28)
 - Port `region.py` (`region.rs`) in [#80](https://github.com/wizard-28/wealthy/pull/80) from [@wizard-28](https://github.com/wizard-28)
 - Port `filesize.py` (`filesize.rs`) in [#83](https://github.com/wizard-28/wealthy/pull/83) from [@wizard-28](https://github.com/wizard-28)
+- Port `cells.py` (`cells.rs`) in [#83](https://github.com/wizard-28/wealthy/pull/89) from [@wizard-28](https://github.com/wizard-28)